Working .py file with edited test file
